requirements.txt
Pillow 
sv_ttk 
torch 
numpy 
